Executor status updating

+++ SchedulerWebApplication/Mutations/ExecutorStatusInput.cs
@@ -4,6 +4,6 @@
     {
         public int ExecutorId { get; set; }
         public ExecutorStatusCode StatusCode { get; set; }
-        public int Date { get; set; }
+        public long Date { get; set; }
     }
 }
